requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
imageio==2.34.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
tifffile==2024.5.22
